Revert "fix (GA-3405): Vendor management  (#279)"
This reverts commit 97ff7e9e7e2148a83772fe6124ca206400952433.

+++ app/controllers/admins/events/station_items_controller.rb
@@ -1,18 +1,21 @@
-class Admins::Events::Stations::StationItemsController < Admins::Events::BaseController
-  before_action :set_station
-  before_action :set_item, only: %i[update destroy]
+class Admins::Events::StationItemsController < Admins::Events::BaseController
+  before_action :set_item_class, except: [:find_product]
 
   def create
+    params[:station_product][:product_id] = @current_event.products.find_or_create_by(name: params[:station_product][:product_name]).id if params[:item_type].eql?("station_product") # rubocop:disable Metrics/LineLength
+
+    @station = @current_event.stations.find(params[:station_id])
     @group = @station.group
     @items = @station.all_station_items
     @items.sort_by!(&@items.first.class.sort_column) if @items.first
-    @item = @klass.new(permitted_params.merge(station: @station))
-    authorize(@item)
+    @item = @klass.new(permitted_params)
 
+    authorize @item
     if @item.save
-      redirect_to [:admins, @current_event, @station], notice: t("alerts.created")
+      redirect_to [:admins, @current_event, @item.station], notice: t("alerts.created")
     else
       flash.now[:alert] = t("alerts.error")
+      @item.product_validations if @item.is_a?(StationProduct)
       @transactions = []
       @devices = []
       @operators = 0
@@ -21,6 +24,8 @@ class Admins::Events::Stations::StationItemsController < Admins::Events::BaseCon
   end
 
   def update
+    @item = @klass.find(params[:id])
+    authorize @item
     respond_to do |format|
       if @item.update(params.require(params[:item_type]).permit(:price, :amount, :hidden))
         format.json { render status: :ok, json: @item }
@@ -31,6 +36,9 @@ class Admins::Events::Stations::StationItemsController < Admins::Events::BaseCon
   end
 
   def destroy
+    @item = @klass.find(params[:id])
+    authorize @item
+
     if @item.destroy
       flash[:notice] = t("alerts.destroyed")
     else
@@ -39,26 +47,25 @@ class Admins::Events::Stations::StationItemsController < Admins::Events::BaseCon
     redirect_to [:admins, @current_event, @item.station]
   end
 
+  def find_product
+    skip_authorization
+    @products = @current_event.products.with_name_like(params[:product_name]).order(:name)
+    render partial: "product_results"
+  end
+
   def sort
     skip_authorization
     params[:order].to_unsafe_h.each_pair do |_key, value|
       next unless value[:id]
       @klass.find(value[:id]).update(position: value[:position].to_i - 1)
     end
-
     head(:ok)
   end
 
   private
 
-  def set_station
-    @station = @current_event.stations.find(params[:station_id])
+  def set_item_class
     @klass = params[:item_type].camelcase.constantize
-  end
-
-  def set_item
-    @item = @klass.find_by!(station: @station, id: params[:id])
-    authorize @item
   end
 
   def permitted_params
@@ -66,9 +73,11 @@ class Admins::Events::Stations::StationItemsController < Admins::Events::BaseCon
                                               :access_id,
                                               :catalog_item_id,
                                               :price,
-                                              :name,
+                                              :product_id,
+                                              :product_name,
                                               :amount,
                                               :credit_id,
+                                              :station_id,
                                               :hidden,
                                               :position)
   end
